add functionality to calculate SI on sagittal images
The centre of mass of the bladder and the distal end of the prostate are used to draw a line.
The intersection between the line and the boundary of the prostate (top left) is then
used as the second point of the SI dimension. Doesnt work perfectly as the SI line can be quite
short. Can probably create an arc centred on prostate distal end with arc drawn on

 bladder centre of mass and use longest calculate SI